refactor(python): Use Poetry (#47)
* docs: Remove unused section

We don't have a root level requirements.txt or setup.py.

* refactor(python): Use relative imports

This allows us to omit the "endpoints" directory from the AWS Lambda
bundle, and lets the IDE resolve references without tweaking PYTHONPATH.

* ci: Lint all Python code

* fix: Create UUID as part of test

The original code ended up with a string like 'Dataset <function uuid1
at 0x7f23e72ff040>.hex'.

* test: Avoid KeyError when running test

Also include the response in full in the error message.

* chore: Ignore CDK synthesize output

It can be useful to manually generate the CDK output locally to debug
what is going to be uploaded to AWS.

* refactor(python): Use Poetry for package management

In CI, since the packages are now installed in a virtualenv rather than
globally, we have to activate the virtualenv at the start of jobs.

The cattrs package has to be kept back to before version 1.1.0 until we
either upgrade to Python 3.7 or AWS CDK moves to a dependency which
doesn't transitively depend on Python 3.7. Issue linked on the relevant
pyproject.toml line.

This merges the production dependencies of the backend and infra
subdirectories, which is not ideal. Once subproject support
<https://github.com/python-poetry/poetry/issues/2270> arrives we should
pull this apart, maybe keeping only the development and test
dependencies in the root. To mitigate this in the meantime, bundle.bash
pulls out only those dependencies which the Lambda function needs.