feat: Tweak commit hooks (#54)
* feat: Use the default yamllint output format

Should be more readable than the parseable output.

* feat: Replace yamlfmt with pre-commit-hook-yamlfmt

The former simply verified that the files were parseable. The latter
works seamlessly as a pre-commit hook.

* feat: Let Black and isort modify files in pre-commit hook

This should result in a faster turn-around than having to manually
format the files.

* feat: Run linters and formatters via pre-commit

This way we won't have to duplicate pre-commit functionality in the CI
scripts.

* feat: Set up kodiak

Use suggested configuration and approve (but not merge) Dependabot pull
requests automatically.