feat: Randomize test sequence
This can detect accidental dependencies between tests, which makes a
test suite brittle.

The job prints the relevant `--randomly-seed=N` value to be able to
re-run a specific sequence.

Using `$GITHUB_RUN_ID`
<https://docs.github.com/en/free-pro-team@latest/actions/reference/environment-variables#default-environment-variables>
as the seed value ensures each push gets a new seed, but rerunning the
workflow for a single push results in the same seed.