test: Randomise test data (#97)
* test: Set random valid dataset type where irrelevant

It is very common to test functionality which requires more than a
single scalar input. In this case creating a dataset requires a type,
title and owning group. But in most of the tests the actual type is
irrelevant as long as it's one of the two valid types, "IMAGE" or
"RASTER". By choosing one or the other *randomly* (or by specifying
`pytest --randomly-seed=N`) we increase the chances of detecting when
our test suite becomes incomplete.

For example, consider the case where we change the fields required for
image and raster datasets to be different. In that case it is likely
that at least some of the tests will start failing half of the time.
This tells us that the type used to be irrelevant, and it no longer is.
The test will have to change, which is likely to prompt us to think
about the tests in this space:

- Do we need one test per type?
- Can we change the test so that the type remains irrelevant?
- Has this test been superseded by other tests?

* test: Set random dataset title where irrelevant

This marks the title as definitely irrelevant where it's supposed to be,
and therefore implicitly as relevant everywhere else. It also ensures
that we learn about any escaping issues immediately.

* test: Set random dataset owning group where irrelevant

* test: Set random dataset ID where irrelevant