feat: Validate stac files recursively (#149)
* feat: Validate metadata links recursively

* feat: Only validate each file once

Since a failing test would imply infinite recursion, we check the number
of calls to make sure we don't have to wait for a very long time to
reach the recursion limit (~90 seconds for a single test).

* perf: Read local schema files

`jsonschema` by default reads file `$ref`s from the web, relative to the
`$id` of the current document. This seriously slows down validation.

* perf: Set up validator only once

Avoids reading the schema files once per file we validate.