feat: Run mypy as part of pre-commit hook
Do *not* enforce it yet, however. This is just to minimize the setup
once we have fixed all the lints.

`require_serial` ensures that `pre-commit` only runs a single `mypy`
command. This is necessary because `mypy` does recursive checks, so by
default we would get duplicate messages about the same issue.

Based on <https://stackoverflow.com/a/59745587/96588>.