feat: Validate checksum (#64)
* feat: Verify stream multihash

* feat: Pass parameters through for validation

* feat: Return non-zero exit code on checksum mismatch

* feat: Print JSON output when validation succeeds

* feat: Print JSON output when validation fails

* feat: Process single item from database

* feat: Require top-level "content" event property

* feat: Disallow any unknown top level event properties

* feat: Require dataset ID event property

* feat: Require dataset version ID event property

* feat: Require next item event property

* feat: Require next item to be non-negative

* refactor: Simplify tests

* fix: Use correct name for version ID property

* feat: Return next item as first item

* fix: Run test with explicit value

Avoids running with -0, which is equal to 0.

* fix: Use proper event structure from design docs

* feat: Lock in next item value

* feat: Make sure first item is non-negative

* feat: Force first/next item to be multiple of iteration size

* feat: Make sure next item is positive

* feat: Force value of iteration size

* feat: Return lower iteration size for partial last iteration

* refactor: Rename variables for clarity

* refactor: Reorder dict

* fix: Use environment-dependent table name

* refactor: Rename constant for clarity

* fix: Move content iterator results to top level in response

* fix: Allow content iterator to describe processing assets table

* fix: Make content iterator return values strings

This is required by AWS. The closest thing to a reference I could find
was <https://forums.aws.amazon.com/thread.jspa?messageID=891811>.

* feat: Increase max iteration size

Batch jobs take a long time to set up and run, so this is a compromise
to make test runs reasonably fast.

* fix: Return minimum iteration size of 2

* fix: Hard-code batch job array size

Works around <https://stackoverflow.com/q/66202138/96588>.

* fix: Give assets table read access from checksum job

* chore: Disable batch job retries

This makes batch jobs much faster to test. We can adjust the retry count
once we've observed how it performs in production.

* feat: Log assets

* test: Add asset file to processing stack test

* feat: Return 0 when asset index does not exist

This is a normal occurrence whenever asset count modulo iteration size
is not zero.

Also, log the error details and parameters in case the issue is not what
we think.

* fix: Return 0 when there is a checksum mismatch

We don't want to terminate the pipeline at that point, so we just log
that it failed.

* refactor: Encapsulate tests which use a logger

Allows importing from the file without creating a logger.

* refactor: Pull out constant for minimum iteration size

And document the min and max values.

* fix: Run database error test on actual infrastructure

It was running internal methods which would also have had to be mocked,
which would be awkward. Also, this ends up testing the actual `cause`
property of the exception, which might work differently from expected.