test: Run only infrastructure tests online
Non-infrastructure tests should not rely on networking, because that
generally means they are interacting with AWS or the STAC spec site when
they shouldn't be.

Once we get any non-infrastructure tests which actually need networking
we'll need to introduce another marker like "networking", but for now
the division is along the existing marker.