fix: Run non-infrastructure tests offline (#371)
* fix: Include STAC core catalog schema in validator

This avoids fetching the JSON files from the STAC spec web site when
running.

* test: Run only infrastructure tests online

Non-infrastructure tests should not rely on networking, because that
generally means they are interacting with AWS or the STAC spec site when
they shouldn't be.

Once we get any non-infrastructure tests which actually need networking
we'll need to introduce another marker like "networking", but for now
the division is along the existing marker.

Co-authored-by: Mitchell Paff <77649372+MitchellPaff@users.noreply.github.com>