test: Remove need for complete storage bucket teardown
This ensures that tests which modify the storage bucket can run in
parallel, while making sure we clean up files afterwards so that we can
tear down the stack.

This does *not* remove all files in case of a test failure. We might
want to make it more robust to deal with that case.