feat: Prep to run tests in parallel (#397)
* test: Verify that each of the IDs are in the response

* test: Use generated dataset IDs

* test: Don't tear down datasets after tests

It's not required for stack teardown, and we use generators for any
identifiers so the chance of a collision should be tiny.

* feat: Don't reset random seed between tests

Resetting it would result in many duplicate generated strings when
running tests.

* test: Remove need for complete storage bucket teardown

This ensures that tests which modify the storage bucket can run in
parallel, while making sure we clean up files afterwards so that we can
tear down the stack.

This does *not* remove all files in case of a test failure. We might
want to make it more robust to deal with that case.

* test: Stop using processing assets table teardown

This interferes with other tests when run in parallel, and isn't
necessary for stack teardown.

* test: Reintroduce single asset dataset version creation test