feat: update dataset catalog with version (#741)
* refactor: Pass dataset prefix earlier in pipeline to avoid

Refactored the input to pipeline to include the dataset prefix.
This is to avoid fetching this from the database multiple times during
the pipeline.

* fix: Remove unnecessary input into pipeline steps

* feat: add new dataset version to dataset catalog

* fix: added tests

* refactor: parametrize tests

* refactor: implement custom layout strategy

* refactor: add test for exception case

* refactor: fix rebase

* refactor: add stronger assertin to update_catalog test

* fix: fix lock file

* fix: fixed broken import from other lambda

* refactor: Strengthen assertions on sqs tests

* fix: Fix inconsistency sqs and lambda casing of message attributes

* fix: Fix table access exception and refactor custom layout strategy

Refactored the input to pipeline to include the dataset prefix.
This is to avoid fetching this from the database multiple times during
the pipeline.

fix: infra

* refactor: Expand usage of STAC custom layout strategy

Added extra test for when adding a dataset version with Item metadata
and expanded the usage of our custom layout strategy to avoid overwriting
existing links when creating a new dataset.

* fix: Add a self ref to root catalog

* refactor: Address PR comments

* refactor: Address futher PR comments