feat: Create and archive JUnit report
Only archive it in case of test failure.

This should allow closer inspection of test result details on failure.

To avoid having to merge two reports, we now run the non-infrastructure
tests *twice,* first to verify that none of them use `socket`, and
second to generate the full test report. This should be fine since they
only take a second or two to run.