fix: Bump Lambda resources (#1456)
* feat: Set Lambda max RAM to 256 MB

This is double the default. All of the jobs were using more than half
their max RAM even for a simple import, so this might speed things up a
bit.

* feat: Set import asset file Lambda max RAM to 512 MB

This is quadruple the default, and should hopefully be enough to import
big files.

* fix: Increase asset file import timeout to max

This job has to wait for potentially large files to be copied across, so
we might need a good chunk of time.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>