feat: Remove concept of Dataset STAC catalog in geostore (#1577)
* refactor: Add a wait for child link to be added to catalog

* refactor: Rename variable and add timeout for wait

* fix: Correct assertions for links in uploaded catalog

* refactor: Align processing test to real world scenario

Calling the dataset endpoint at the beginning of the processing stack test will trigger the creating of STAC catalogs and so create a more real test scenario.

* refactor: address PR comments

* feat: Remove concept of Dataset STAC catalog in geostore

Instead, new versions will be referenced directly in the root catalog. This is an intermediary step to allowing Datasets to be updated in place rather than having to create a new dataset version.

* fix: remove unnecessary imports

* refactor: update processing tests to no longer reference dataset catalog