refactor: Raise NotImplementedError
The get_item_href function isn't called post normalized_href removal; however, it is needed to override all the abstract methods in the parent class before it can be instantiated. raise NotImplementedError is added to ensure pragma: no cover is removed if this method is called again in the future as the codebase expand.