refactor: Better logging when eTag is unavailable
When comparing etag to check if object is available and identical before writing to bucket, it is likely that most of the response is no such key (since a significant write request is uploading object to the bucket the first time around). There is no point logging this since we are only interested in errors other than 404 (e.g. 403 permission).