refactor: Better tests for getting etag
We should be testing how etag is calculated and returned from aws by uploading random file(s) to s3 bucket. This should mimic the actual behaviour of uploading an object to s3 and calculating the etag locally before comparing their values. This is much better than pre-specifying expected etag values and comparing them locally, as it may not necessarily mimic the actual behaviour on aws.