fix: Remove workflow schedules (#2061)
* build: Update copyright years manually for now

Works around <https://github.com/orgs/community/discussions/32197>.

* fix: Remove workflow schedules

GitHub unfortunately disables workflows with a schedule on any repo without
activity for 60 days. This disabling is silent and invisible, only showing
up as a slightly different icon next to the job on the Actions page.
Clicking that icon will prompt whether you want to re-enable it. But at
that point the damage is already done, in shape of a lot of fiddling to
figure out what went wrong, and having to re-push PRs with different
contents to actually trigger the now active workflow.

There's been no response from GitHub about this issue yet
<https://github.com/orgs/community/discussions/32197>.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>