feat: Compare s3 object etag in S3StacIO (#2029)
* feat: Check if object exists and is identical in s3 before uploading

Pystac .save method (re)writes metadata when called, regardless if an identical metadata already exists. This commit checks the eTag of the object in s3 (if exists) before (re)writing the object.

* feat: Add unit tests to check s3 etag

* refactor: implement suggestions made in pr #2029

* fix: Allow lambda to access param store

geostore_s3_client in s3_utils.py needs s3 client role to query s3 object etag

* refactor: Better logging when eTag is unavailable

When comparing etag to check if object is available and identical before writing to bucket, it is likely that most of the response is no such key (since a significant write request is uploading object to the bucket the first time around). There is no point logging this since we are only interested in errors other than 404 (e.g. 403 permission).

* refactor: Better logging when eTag is unavailable

When comparing etag to check if object is available and identical before writing to bucket, it is likely that most of the response is no such key (since a significant write request is uploading object to the bucket the first time around). There is no point logging this since we are only interested in errors other than 404 (e.g. 403 permission).

* refactor: Call head_object instead of get_object

We are only interested in s3 object metadata. It makes sense to call head_object.

* fix: Chunk size shouldn't be configurable if using aws default value

This value should be consistent throughout geostore.

* refactor: Better tests for getting etag

We should be testing how etag is calculated and returned from aws by uploading random file(s) to s3 bucket. This should mimic the actual behaviour of uploading an object to s3 and calculating the etag locally before comparing their values. This is much better than pre-specifying expected etag values and comparing them locally, as it may not necessarily mimic the actual behaviour on aws.

* refactor: tidy up

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>