feat: Delete CloudWatch Log Groups over a week old
There is no mechanism in place to clean up CloudWatch Log Groups. While the logs itself can be set to expire, the Log Groups are kept forever. There isn't an elegant way to delete log groups within CDK on stack teardown currently (this may change in the future - see https://github.com/aws/aws-cdk/issues/21804). This is a temporary workaround until a better solution is found.

Note: A --no-paginate option is added to limit the number of queries (~50). This prevents rate limiting as well as stops a potentially endless number of results being returned by AWS (which would result in a very long-running workflow). This means only a maximum number of 50 log groups are removed each workflow run.