fix: Delete Cloudwatch log groups race condition
It is possible that the delete-log-group command to gets into a race condition when more than one CI runs at the same time (i.e. it tries to delete the same log group that another CI has just deleted moments ago). This causes xargs to throw a 123 exit code, which, in turn causes GitHub actions to fail. Ignoring this exit code is one way to get around this problem as a workaround. It should be noted that this implementation prevents other xargs errors with the same exit code from being raised.