feat: Check paths (#2589)
* refactor: Replace space with hyphen in filenames

For ease of handling.

* feat: Check file names

Verifies that file names don't contain any characters outside the POSIX
portable file name character set:

> ASCII letters and digits, ‘.’, ‘_’, ‘-’, and ‘/’.

and that they don't start with a hyphen.

The exclude is in place to ignore one of the checks, the 14-character
POSIX max path component length, while still enforcing the other checks.
This might be possible to remove in the future, if
<https://lists.gnu.org/archive/html/coreutils/2023-01/msg00006.html> is
implemented.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>