refactor: Use class variable instead of lru_cache (#2600)
lru_cache is a bit of a hacky workaround to prevent values from being updated within the context of setting up a lambda_packaging_dir using mkdtemp. This isn't how it is intended to be used. A class variable would provide the same outcome whilst keeping the code a bit tidier.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>